Fixed name sanitation / inconsistent filenames
This change solves an issues that the file path to be read from
/ written to is sometimes not the same as the actual path that already
exists.

Added a "sanitize_name" function for easy and consistent sanitation.
This function is now used by scraper.py and get_book_pretty_filename to
sanitize the author and book names.

This also affects the get_book_pretty_filepath because it uses the
get_book_pretty_filename function to figure out the correct path.