Book Downloading - Continued
The procces of checking if files were downloaded or not already is now
based on a for loop method. Exact name matches are now checked instead
of just the number of *.m4a files present.

Doing it this way is slightly more complicated, but also has the
potential for better accuracy in certain cases.

Also, the console messages for various steps of the downloading proccess
have been simplified in order to use less lines, be more consistent in
wording, and also be more clear about what is going on.

Example:
[.] Audio for all 9 blinks already exists, skipping download...
[.] Combining, and keeping originals of, audio files for 5-levels-of-lea
dership-en

Instead of:
[.] Found 9 out of 9
[.] Audio is already downloaded. Skipping.
[.] Combining audio files for 5-levels-of-leadership-en
[.] Keeping audio files.