Changed requirements.txt to minimum version, only
The version number listed for lxml is not the newest one available (for
Python 3.9). The "==" operator forced lxml to use version 4.5.2, even if
that meant downgrading it. This causes errors for lxml if "Microsoft
Visual C++ 14.0" was not installed (because it is needed for compiling).

Also, some packages like "certifi" always needs to be up to date (for
security sake).

In response to these issues, all of the version numbers should be
treated as minimum version numbers (instead of required ones).